requirements.txt
numpy==1.24.3
matplotlib==3.7.1
pandas==2.0.3
opencv-python-headless==4.8.0.74
streamlit==1.25.0
Pillow==9.5.0